feature / order,orderItem Entity id 타입 변경

+++ src/main/java/com/github/secondproject/order/entity/OrderItem.java
@@ -15,7 +15,7 @@ public class OrderItem {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "orderItem_id")
-    private Integer orderItemId;
+    private Long orderItemId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "order_id", nullable = false)
